+++ ui-spa/playwright/integration-tests/activity-log.spec.ts
@@ -1,7 +1,7 @@
 import { delay, HttpResponse, http } from "msw";
 import { expect, test } from "./utils/test";
 import { Page } from "@playwright/test";
-import { ActivityLogResponse } from "../../src/schemas/responses/activityLogResponse";
+import { type ActivityLogResponse } from "../../src/schemas";
 test.describe("activity log", () => {
   const goToActivityLog = async (page: Page) => {
     await page.goto("/case/12/case-management");
