+++ ui-spa/playwright/integration-tests/transfer-error-page.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "./utils/test";
 import { delay, HttpResponse, http } from "msw";
-import { TransferStatusResponse } from "../../src/schemas/responses/transferStatusResponse";
+import { type TransferStatusResponse } from "../../src/schemas";
 import { Page } from "@playwright/test";
 
 test.describe("transfer-error-page", () => {
