+++ ui-spa/src/common/utils/getFormattedEgressFolderData.test.ts
@@ -1,5 +1,5 @@
 import { getFormattedEgressFolderData } from "./getFormattedEgressFolderData";
-import { EgressFolderData } from "../../schemas/responses/egressFolderData";
+import { type EgressFolderData } from "../../schemas";
 
 describe("getFormattedEgressFolderData", () => {
   it("Should return the formatted egress folder data with new path which appends file and folder name to the end", () => {
