+++ ui-spa/src/common/utils/getMappedResolvePathFiles.ts
@@ -1,4 +1,4 @@
-import { IndexingError } from "../../schemas/responses/indexingFileTransferResponse";
+import { type IndexingError } from "../../schemas";
 import { getRelativePathFromPath } from "./getRelativePathFromPath";
 import { getFileNameFromPath } from "./getFileNameFromPath";
 import { ResolvePathFileType } from "./getGroupedResolvePaths";
