+++ ui-spa/src/components/egress-connect/EgressSearchResults.tsx
@@ -1,6 +1,5 @@
 import { useMemo, useState } from "react";
-
-import { EgressSearchResultData } from "../../schemas/responses/egressSearchResult";
+import { type EgressSearchResultData } from "../../schemas";
 import { Button, SortableTable, Tag } from "../govuk";
 import { formatDate } from "../../common/utils/formatDate";
 import { UseApiResult } from "../../common/hooks/useApi";
