+++ ui-spa/src/mocks/data/caseMetaData.dev.ts
@@ -1,4 +1,4 @@
-import { CaseMetaDataResponse } from "../../schemas/responses/caseMetaData";
+import { type CaseMetaDataResponse } from "../../schemas";
 export const caseMetaDataDev: CaseMetaDataResponse = {
   caseId: "12",
   egressWorkspaceId: "egress_1",
