+++ ui-spa/src/mocks/data/egressFolderResults.playwright.ts
@@ -1,4 +1,4 @@
-import { EgressFolderResponse } from "../../schemas/responses/egressFolderData";
+import { type EgressFolderResponse } from "../../schemas";
 export const egressRootFolderResultsPlaywright: EgressFolderResponse = {
   data: [
     {
