+++ ui-spa/src/mocks/data/transferStatus.dev.ts
@@ -1,4 +1,4 @@
-import { TransferStatusResponse } from "../../schemas/responses/transferStatusResponse";
+import { type TransferStatusResponse } from "../../schemas";
 export const egressToNetAppTransferStatusDev: TransferStatusResponse = {
   status: "Completed",
   transferType: "Copy",
