+++ ui-spa/src/mocks/data/transferStatus.playwright.ts
@@ -1,4 +1,4 @@
-import { TransferStatusResponse } from "../../schemas/responses/transferStatusResponse";
+import { type TransferStatusResponse } from "../../schemas";
 export const egressToNetAppTransferStatusPlaywright: TransferStatusResponse = {
   status: "Completed",
   transferType: "Copy",
